requirements.txt
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.1
azure-functions==1.12.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==21.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coverage==6.5.0
cryptography==38.0.3
exceptiongroup==1.0.4
grpcio==1.50.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
packaging==21.3
pluggy==1.0.0
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-mock==3.10.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
